[semantic] remove allowed warnings
